containers/camera/requirements.txt
matplotlib==3.8.2
numpy==1.25.2
omegaconf==2.3.0
onnxruntime==1.19.2
opencv-python==4.9.0.80
openpyxl==3.1.1
packaging==23.0
pandas==1.4.1
pillow==11.0.0
pillow_heif==0.20.0
pipenv==2023.9.1
plotly==5.11.0
pydantic==2.9.2
pydantic_core==2.23.4
regex==2022.10.31
sahi==0.11.18
scikit-learn==1.2.2
scipy==1.9.3
seaborn==0.13.2
timm==1.0.8
torch==2.4.0
torchvision==0.19.0
tqdm==4.66.5
transformers==4.44.2
ultralytics==8.3.29
ultralytics-thop==2.0.11
fastapi==0.110.0
uvicorn==0.27.1
starlette==0.36.3